Version 1

+++ src/main/java/com/application/data/entity/TicketStatus.java
@@ -4,7 +4,7 @@ import jakarta.persistence.Entity;
 
 @Entity
 public class TicketStatus extends AbstractEntity {
-    private Long statusId;
+    private Integer statusId;
     private String statusName;
 
     public TicketStatus() {
@@ -14,16 +14,16 @@ public class TicketStatus extends AbstractEntity {
     public TicketStatus(String statusName) {
         this.statusName = statusName;
     }
-    public TicketStatus(Long id,String statusName) {
+    public TicketStatus(Integer id,String statusName) {
         this.statusId = id;
         this.statusName = statusName;
     }
 
-    public Long getStatusId() {
+    public Integer getStatusId() {
         return statusId;
     }
 
-    public void setStatusId(Long statusId) {
+    public void setStatusId(Integer statusId) {
         this.statusId = statusId;
     }
